+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Borderless Topic view for more characters of topic names
 - Display broker text easier to read
+- Display version & broker in the lower right corner
+- Display MQTT connection error in its own area
 
 ## [0.19.0] - 2023-05-17